+++ dio/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Remove the force conversion for the response body.
 - Fix `DioErrorType.cancel` in `Interceptors`.
 - Fix wrong encoding of collection query parameters.
 - Fix "unsupported operation" error on web platform.